Fix ETP local iptables rules for host access
When pod is trying to access nodeport, it will now go to the node's mp0.
From there, for ETP local traffic, the packet will be DNAT'ed to the
masquerade IP 169.254.169.3, and routed back into mp0. This is where it
would hit the special per node LB that would handle this traffic and
send it to the local pod.

In normal LGW mode this traffic skips SNAT via another iptables rule so
that the traffic will be delivered to the pod without source IP
modification. However, for pod -> nodeport, traditionally we have always
treated it as egress traffic and SNAT'ed to the node IP. Now that we
route to mp0, there is no more SNAT.

Therefore when the packet arrives at the node's mp0, and it is sent back
into OVN without SNAT, towards the local endpoint, the reverse path will
be asymmetric and go straight back to the original pod which breaks the
flow. This patch, ensures we SNAT only pod traffic accessing nodeport
that is being hairpinned back into mp0.

Signed-off-by: Tim Rozet <trozet@redhat.com>