Observability: change decode funcs to provide structured data
This change allows consumers to get structured data instead of pre-built
strings. A use-case is if consumers want to produce metrics out of the
decoded events, without having to parse the string to extract meaningful
data such as the network policy name or direction.

The previous string representation is still accessible by calling `String()`
method from the resulting objects.

Signed-off-by: Joel Takvorian <jtakvori@redhat.com>