podNetworkController: Switch enabled LSP flags for VM migration
Currently all pods have the LSP.Enabled option set to true by default.
During migration, in order to prevent unneeded packet loss the
lsp.Enabled of the migration pods (source and target) need to be as
follows:
1. until target pod is domain ready: source pod lsp is enabled, target
pod lsp is disabled.
2. when target pod is domain-ready: source pod lsp is disabled, target
pod lsp is enabled.

This switch significantly reduces the migration downtime.
Currently it is restricted to VM with only secondary interfaces, and
layer2 topology.

Signed-off-by: Ram Lavi <ralavi@redhat.com>