podNetworkController: Switch back source pod LSP after failed migration
Currently during migration there is a switch in the LSP flags between
the source and target pods, once the target pod is domain-ready.
But in case the where the migration failed (=target pod is
failed/completed and the source pod is still alive) after the target-pod
was ready then the switch needs to be reverted.

This corner case is handled in the DeleteResource() code path.

Signed-off-by: Ram Lavi <ralavi@redhat.com>