Refactor nftables add/delete handling
Rename nodenft.UpdateNFTElements to nodenft.AddObjects, and
nodenft.DeleteElements to nodenft.DeleteObjects.

Make DeleteObjects deal with the case of (reliably) deleting map
elements with unknown values by having it do a List before the Deletes
to ensure it only deletes the elements that already exist.

Also, push "context" handling back out to the callers to get rid of
`context.TODO()`.

(Partly in preparation for making this more complicated when I port
over generateIPTRulesForLoadBalancersWithoutNodePorts, which will
require support for syncing Rules in addition to Elements.)

Signed-off-by: Dan Winship <danwinship@redhat.com>