kubevirt, pudn: Keep SNAT after live migration
After kubevirt live migration the migration source pod is deleted and
this end up deleting the per pod snat that allow north south traffic on
that node also the migrated VM has still the default gw mac address
pointing to the previous node and not to the node where the VM is
running, that get updated after ip neigh reconciliation and that depends
on the guest. This end up with VMs not being able to do north south
traffic after live migration.

This change keep the SNAT at all he nodes where the VM is being migrated
from so north south traffic works even if the VM default gw mac is
pointing to previous nodes.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>