Allow default network pods to reach localnet on the same node
Add a flow at prio=101 to allow traffic from the OVN network to the host network to be handled normally by the bridge instead of forwarding it directly to the NIC (which would be done by the flow at prio=100). This allows pods in the default network to reach localnet pods on the same node.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>