Add flow for localnet -> host on same node
When a localnet is mapped to breth0, localnet pods were already able to reach the host network as long as the destination was on a different node. Let's make the behaviour consistent and allow them to reach also the host network on the same node, for instance a host-networked pod on the same node.

In terms of traffic paths, this enables host->host on the same node.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>