E2E: add EIP feature check when "PodNetwork" is advertised
This PR attempts to add EIP coverage for when CDN is advertised.
When CDN is advertised and a pod is selected for EIP, src IP is
EIP. If it is not selected, src IP is pod IP.

Our current external "target" containers do not contain routes
for CDN therefore cannot route reply traffic.

However, we can utilize an existing job that sets up resources.
EIP tests can use them if they are present.

Ideally, Id like to CRUD those resources within the test itself
and remove the dep on external config, however I want to opt into
the existing BGP lane upstream and this is the easiest path.

Currently, we have a CI job which can create a ext FRR+Server
and we advertise CDN, therefore Server contains a default route
towards the FRR which contains the necessary routes back to the
k8 node. Server is agnhost which may return source IP. We also
use httpd and we can see within the logs of this app what source
IP it saw.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>