Fixes a nodePort service targeting a pod with hostNetwork e2e
The test case never really worked as we thought. It was working under a
premise that sending UDP packets via netcat would trigger ICMP needs
frag. This does not happen in local gateway mode, because the DF bit on
the packet is not set to 1. So for local gateway mode, packets will be
fragmented as they hit the nodeport on the intermediary node.

For shared gateway mode, packets to a host networked pod endpoint are
not fragmented and are rerouted out of GR back to the next host endpoint.
However, for OVN networked endpoints and shared gateway mode, the packet
will hit the GR and then even if DF bit is not set, OVN GR will send
back an ICMP needs frag (because OVS is incapable of fragmentation
across different MTU boundary interfaces).

Since our new logic prevents ICMP needs frag from installing a cached
MTU route, the client will continue to send packets with a 1500 byte MTU
and they will be dropped. We choose not to fix this for now as it was
identified as not a practical use case:
https://issues.redhat.com/browse/OCPBUGS-7609

It could be fixed in the future by using an ip rule to match on nodeport
port range, and then redirecting traffic to another routing table with a
lower MTU on the route.

Signed-off-by: Tim Rozet <trozet@redhat.com>