Update e2e pod2pod-tcp-low-mtu MTU handling
This test was originally designed to somewhat simulate a deployment
where some workers were on different subnets than others, with different
MTU values. The test would see a pod try to send a packet that was
larger than the MTU acceptable by Geneve to the other node, and ICMP
needs frag would be triggered and sent back to the pod.

For KIND we have no way of deploying workers in different subnets, so
for now as a poor mans way to exercise the test, we artificially lower
the MTU of the nodes and then revert it back after the test.

Signed-off-by: Tim Rozet <trozet@redhat.com>