Refactor routemanager
For a number of reasons:

1. Performance

With BGP the number of routes that can be installed on the system scales
up. Tests with > 1500 routes causes trouble. The current sync
implementation has exponential loops. On each iteration,
RouteListFiltered is called which dumps all the routes on the system and
then filters client side. netlink API is by value, and this causes high
GC crunch.

Changed sync implementation to have sequential loops and a single list
call to netlink. Changed to avoid listing routes in any other case.

sinc is still a heavy operation that keeps a lock and migh require
further optimizations. This is a first take.

2. Consistency

The used RoutePartiallyEqual had a different criteria than what was used
to actually delete a route from netlink, leaving the door open for
inconsistencies. This could be thought of as a compromise due to the fact
that routes can assume different defaults when installed than those
provided. Proposal is to do a bit better and ignore those defaults when
comparing against kernel routes as a unified criteria. Should fulfill our
use cases but might need further tweaking in the future. Potential
problem is that kernel switches bad an incorrect value to default
instead of throwing us an error which can cause a route to constantly
sync. Much of this is undocumented as far as I could tell but anyway has
not been observed with the specific routes that we install.

Also, the current implementation was validating and indexing by link index
which is actually not a required field of a route. Instead, key routes by
priority, prefix and table. Route manager assumes ownership of all the
routes with any of the provided keys and will delete unrecognized routes
on sync. This is inline to what our previous netlink delete was doing.

3. Other

* Reset ticker on each sync iteration to avoid backlogging.
* Cleaned up excessive logging

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>