[db sync] remove db sync from the older version.
ACL, address_set and port_group sync functions are only needed as long
as the previous ovn-k release has the old db format. In the last ovn-k
release 1.0.0 this format is not present, so it is safe to remove these
functions.

When ACL tiers were introduced, all ACLs have got a default tier value
of Primary tier (=0). To move all ACLs to the Default tier, a part of
the ACL sync function was introduced that updates Primary ACL tier to
the Default for all ACLs. This part has to be removed now when we are
starting to use Primary ACL tier. The rest of the cleanup is not
strictly required, but good to have while we are here.

Signed-off-by: Nadia Pinaeva <npinaeva@redhat.com>