allocator/pod_annotation: Setting hard-coded MAC for pod
Since the IP is not assigned, and since in primary-udn case there is no
MAC/IP request annotation, the MAC is randomly generated in a loop.
In order to prevent this, add an annotation to force the MAC assignment
to be hard-coded to 00:AA:BB:CC:DD:EE.
In future commits, other modifications will be needed in order to block
l3 configurations of the pod. For that intended cause and for the sake
of the POC only - all future commits will assume this this MAC as an
indicator to the poc-pod.

Signed-off-by: Ram Lavi <ralavi@redhat.com>