util/pod_annotation: Force ip related params to pod-annotations ovn annot
Since IP dynamic assignment is removed, the ovn's pod-networks
annotation is not added with the gateway and routes (becuase it is added
per IP and now we have no IP).
In order to keep this information on the pod-networks annotation, and
since the data only needs the subnet in order to determine the above
params, it is set manually.
Also, the podAnnotation's "gateway_ip" is manually set to match the
value of a "classic" primary-udn pod.

For the sake of the POC only - this is done only in case the agreed upon
MAC (00:AA:BB:CC:DD:EE) is assigned, and the submet is 10.0.0.0/24.

Signed-off-by: Ram Lavi <ralavi@redhat.com>