cni/helper_linux: Skip network setup
Since IP dynamic assignment is removed, the network's link has no ip.
This causes all route, and gateway settings using netlink to fail (cause
linux expects the ip to be reachable when the route is set).
Therefor the routes and gateways rules are skipped, assuming this should
be set manually by the user via cloud-init.

For the sake of the POC only - this is done only in case the agreed upon
MAC (00:AA:BB:CC:DD:EE) is assigned.

Signed-off-by: Ram Lavi <ralavi@redhat.com>