cni/cni.go: Mimic ip entry in cni result
in a VM connected via primary UDN, in order for the virt-launcher pod to
properly connect the bridge to the vm guest, it depends on the
network-status annotation. Specifically, it needs to have the ovn-udn1
interface set as "default: true", otherwise it connects the pod to the
cluster network "eth0" interface [0][1].
However, Since IP dynamic assignment is removed, the network's link has
no ip, ergo - the cni.result will not have the cni.result.ips entry with
the ovn-udn1 interface.

This commit manually adds the ovn-udn1 ip entry to the cni.result in
order for the VM to properly connect to the guest.
Note that the cni.result.ips[*].Address.ip must not be empty, or
multus-cni will fail unmarshaling the result [2]. Instead, it is set as
"0.0.0.0/0".

As an alternative to using "0.0.0.0/0", a unit test is added to propose
a new custom Unmarshal method for Response object. If selected, this new
Unmarshal method should replace the one set on multus-cni [2].

[0]
https://github.com/kubevirt/kubevirt/blob/0878ec6fdb01b8e1a14e6a49ce7f9177b204544c/pkg/network/controllers/vmi.go#L66
[1]
https://github.com/kubevirt/kubevirt/blob/e36bb0bd799764901e5dade8e4b2a5e906230d15/pkg/network/multus/status.go#L55
[2]
https://github.com/k8snetworkplumbingwg/multus-cni/blob/master/pkg/server/api/shim.go#L119

Signed-off-by: Ram Lavi <ralavi@redhat.com>