Drop in_port from ip dispatch OF rule
The in_port was to match on IP traffic coming from the physical link to
be dispatched to conntrack and table 1 to find out whether the packet
was a reply to the host or to OVN. We are now conntracking these packets
also as they go to localnet ports attached to the bridge. Therefore we
need to also match on packets from those ports.

We do not want traffic from OVN or from LOCAL to hit this flow, but that
should be avoided by higher priority flows.

Signed-off-by: Tim Rozet <trozet@redhat.com>