ovnkube.sh: use node name as zone as default if ovn-ic
If the `k8s.ovn.org/zone-name` label is not set on the node,
the fallback logic now uses the node name as the zone when
`ovn_enable_interconnect` is true. Otherwise, the zone
defaults to "global" as before.

Also updated the empty string check to use `-z`, which is
more idiomatic in Bash.

Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>