Fix LB service e2e test
The test "Load Balancer Service Tests with MetalLB Should ensure
connectivity works on an external service when mtu changes in
intermediate node" relied on forwarding traffic through localhost into
the cluster docker kind network. I was able to reproduce it locally and
looking into filter rules, that traffic is dropped and requires an
additional rule to allow it.

I do not know what change triggered the failure. It looks like it might
have been a docker update in the github runner image [1].

[1] https://github.com/actions/runner-images/commit/ce61c11aaca04737a02288ec2ad5cc0c3b2d74c3

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>