contrib, kind.sh: Fix local registry when using podman
On podman push, it defaults to secure connection.
In our case the local registry uses an insecure connection result in podman push failures making it impossible to work with the local registry when podman is installed.

Set podman to skip secure connection check when pushing OVN-K images to the local registry.

Signed-off-by: Or Mergi <ormergi@redhat.com>