ovspinning: detect reservedSystemCPUs
When https://github.com/kubernetes/kubernetes/pull/127483 is enabled on the cluster
ovnkube-node pod is not aware of the reservedSystemCPUs which means
the affinity mask of ovsVSwitchd and ovsdbserver processes
will not include the reserved CPUs anymore.

This might affect the above processes performance.

This commit contains the logic to discover the reservedSystemCPUs
of the node and add them to the affinity mask.

In case the reservedSystemCPUs is empty, it will keep
the existing affinity mask.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>