IC: gate ovn-controller start on ovnkube controller syncing
For EIP, when a Node is assigned an EIP and if the EIP moves
when ovnkube-controller is down and SB DB contains stale SNAT
entries, then when ovn-controller starts again, it will GARP.
Ideally, ovn would have support for suppressing GARPs for LRPs,
but that is not available.

I took a more general solution to fix the issue. For IC, gate
ovn-controller until SB DB is not stale.

The previous implementation of ovnkube-controller sync+startup
depended on ovn-controller running. I removed this dependency.
ovnkube controller writes a file when startup is complete and
changes flow to SB DB.

External entities can predicate on this file to determine if
the ovnkube controller has sync.
ovnkube controller will remove this file on exit.

ovn-controller start is now gated on the presence of this file
for IC. We do not support multi Node per Zone therefore ovnkube
controller is co-located on the same name as ovn-controller.

If theres a pod delete and recreate, ovn-controller will now
not start until SB DB is not stale. Before, we accepted this
and got incremental updates during the ovnkube controllers sync+start.

if ovn-controller container restarts and ovnkube controller did not,
ovn-controller starts because the SB DB hot file exists.

if ovnkube-controller restarts and ovn-controller did not, then
ovn-controller will continue running.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>