fix intermittent disk space issue
sometimes a tmp pv image file used for setting up the runner
gets leftover in /mnt. Add the removal of that in the "Free
up disk space" step.

Also move that step to a composite action to deduplicate it
for easier maintenance

Also added the Runner Diagnostics step as the first step
that runs for every job

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>