Switch to netlinksafe wraps for a limited set of funcs from pkg vishvandanda/netlink
..that may return EINTR.

Fixes: https://github.com/ovn-kubernetes/ovn-kubernetes/issues/5358

A change to the vishvananda/netlink package (vishvananda/netlink#925 in v1.2.1)
exposes NLM_F_DUMP_INTR in some netlink responses as an EINTR error return but
also doesn't return any data therefore all we can do is retry and netlink
safe provides this with wrappers to certain netlink fn.

This has been causing failures in some unit tests that interact with netlink.

Retry the requests on EINTR, up to five times,

Also, add RuleList to the the API and switch IP rule manager to use this.
The reason it wasnt done before is ip rule manager doesn't fake out
netlink for testing.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>