pod_annotations: Block duplicate IP on new pods
Currently when a pod is reconciled and IPs are requested, the
ipAllocator's "already allocated" errors are ignored.
In order to block IP conflicts with other pods, only ignore duplicate IP
errors in cases when PreconfiguredUDNAddressesEnabled FG is enabled and:
- the NAD's network annotation has already persisted, indicating that
the pod was already reconciled and assigned with IPs.
- the pod is a VM-pod that is during an active migration process, when
two pods under the same VM temporarily hold the same IPs.

Under these new terms, when ErrAllocated is discovered and not ignored -
publish this as a pod event.

Signed-off-by: Ram Lavi <ralavi@redhat.com>