Only configure OVN conditional SNATs when there are pods already.
This ensures that we only keep the conditional snat if the network
already existed and pods are already wired. During an upgrade, the node
will be drained and pods will be moved. When OVNK comes up at that point
we no longer have pods on the node, and now we can remove this SNAT in
OVN, and use the one in the host nftables.

Signed-off-by: Nadia Pinaeva <n.m.pinaeva@gmail.com>
Co-authored-by: Tim Rozet <trozet@redhat.com>