E2E: add EIP TCP disruption test where client pushes externally
Client sets up tcp test with external server and performs
an iperf test while we restart control plane.
Client contineously pushes data externally.

Fails on dataloss.
Logs retransmissions.
Restarts ovnkube node on egress node and client pod node.

This test cannot work currently with the gRPC
implementation of EIP Healthcheck because the EIP is prone
to move to a different Node and therefore that
will interrupt the tcp connection.
The gRPC server which helps implement the Egress IP
healthcheck is removed when ovnkube node is shutdown / restarted,
and there's a possibility the node is no longer assigned as an
egress node by cluster manager, and today that will break
the TCP connection if the eip is unassigned or reassigned
to another node switch back to gRPC when [1] is merged.

[1] https://github.com/ovn-kubernetes/ovn-kubernetes/pull/4286

Signed-off-by: Martin Kennelly <mkennell@redhat.com>