Openflow: Reply traffic destined to OVN should have mac dst breth0
In openflow table 1, at priority 100 we match reply traffic and dispatch it to either OVN or the host according to the packet mark we see.

Now, there are cases in which reply traffic destined to OVN actually goes through an external router (e.g. localnet pods replying to pods in the default network through an external router). In this case, we should distinguish between reply traffic whose next hop is OVN/host and reply traffic whose next hop is elsewhere. Let's add dl_dst=breth0 to the prio=100 flows in table 1 in order to do that. Let's also modify accordingly the flows in table 11, where traffic is fragmented when necessary, and if not, it's just output to OVN.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>