EIP: do not skip generating potential valid next hops if LSP not found
For IC mode, there is no expectation we can fetch
a remote nodes LSP, therefore, by skipping (continue),
it is causing us to skip generating valid next
hops for the remote node.

Later in sync LRPs, when a valid next hop is inspected,
we do not find it valid and remove that valid next hop.

Handlers will re-add it shortly later.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>