Fix getHostNamespaceAddressesForNode error wrapping
Incorrect wrapping of errors in getHostNamespaceAddressesForNode
prevented callers to assert the type of error returned. In this specific
case the caller delIPFromHostNetworkNamespaceAddrSet could not appropriately
check that the error returned was AnnotationNotSetError and was failing
a deletion flow that should be noop instead.

This was triggered by [1] changes that cause the deletion flow to depend on
annotations set by different controllers instead of annotations set
atomically by a single controller, making it possible that some
annotations were set but not others and opening up a new error handling
flow that wasn't possible before. The scenario is a node that is added
and deleted quickly enough for some controller in cluster manager to be
able to annotate the node and some not.

[1] https://github.com/ovn-kubernetes/ovn-kubernetes/pull/5396

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>